feat: Implement Hamming distance method for vector comparison
This commit introduces a new feature that allows users to compare vectors using the Hamming distance method. The Hamming distance is a measure of the number of differences between two vectors of the same length. This feature can be useful for a variety of applications, such as pattern recognition and error correction.

This commit includes the following changes:

* Added two example vectors, e1 and e2.
* Added a function to calculate the Hamming distance between two vectors.
* Added a loop to iterate over the input vectors and compare them to the example vectors using the Hamming distance.
* Added logic to determine which example vector is most similar to each input vector.

These changes allow users to easily compare vectors using the Hamming distance method. This feature can be used to improve the performance of a variety of applications.